Internals

MicroCollections.dict0 — Function

dict0(T::Type = Pair{Union{},Union{}})

Create an empty dict shim of element type T.

MicroCollections.dict1 — Function

dict1(k => v)

Create a singleton dict with key k and value v.

MicroCollections.set0 — Function

set0(T::Type = Union{})

Create an empty set shim of element type T.

MicroCollections.set1 — Function

set1(x)

Create a singleton set with element x.

MicroCollections.upcast — Function

MicroCollections.upcast(singleton_container::C) -> container::C